Restore Add functionality in VariableController
Add try-catches for on-methods to improve debugging

+++ src/org/daxplore/producer/gui/view/variable/RawVariableTableModel.java
@@ -47,7 +47,7 @@ public class RawVariableTableModel extends DefaultTableModel {
 				if(info.getValue() != null) {
 					int index = ms.getOptionIndex(info.getValue());
 					if(index != -1 && availableToNumbers.contains(index)) {
-						toNumberMap.put((Double)info.getValue(), index);
+						toNumberMap.put(info.getValue(), index);
 					}
 				}
 			}
